+++ src/bin/cargo-deadlinks.rs
@@ -1,5 +1,6 @@
+use std::env;
 use std::path::PathBuf;
-use std::process;
+use std::process::{self, Command};
 
 use cargo_metadata::MetadataCommand;
 use docopt::Docopt;
@@ -19,6 +20,7 @@ Options:
     -h --help               Print this message.
     --dir                   Specify a directory to check (default is target/doc/<package>).
     --check-http            Check 'http' and 'https' scheme links.
+    --no-build              Do not call `cargo doc` before running link checking. By default, deadlinks will call `cargo doc` if `--dir` is not passed.
     --debug                 Use debug output. This option is deprecated; use `RUST_LOG=debug` instead.
     -v --verbose            Use verbose output. This option is deprecated; use `RUST_LOG==info` instead.
     -V --version            Print version info and exit.
@@ -30,10 +32,11 @@ struct MainArgs {
     flag_verbose: bool,
     flag_debug: bool,
     flag_check_http: bool,
+    flag_no_build: bool,
 }
 
-impl From<MainArgs> for CheckContext {
-    fn from(args: MainArgs) -> CheckContext {
+impl From<&MainArgs> for CheckContext {
+    fn from(args: &MainArgs) -> CheckContext {
         CheckContext {
             check_http: args.flag_check_http,
             verbose: args.flag_debug,
@@ -54,17 +57,20 @@ fn main() {
     let dirs = args
         .arg_directory
         .as_ref()
-        .map_or_else(determine_dir, |dir| vec![dir.into()]);
+        .map_or_else(|| determine_dir(args.flag_no_build), |dir| vec![dir.into()]);
 
-    let ctx = CheckContext::from(args);
+    let ctx = CheckContext::from(&args);
     let mut errors = false;
     for dir in dirs {
         let dir = match dir.canonicalize() {
             Ok(dir) => dir,
             Err(_) => {
                 println!("Could not find directory {:?}.", dir);
-                println!();
-                println!("Please run `cargo doc` before running `cargo deadlinks`.");
+                if args.arg_directory.is_none() {
+                    println!();
+                    println!("deadlinks incorrectly guessed the target directory for the documentation.\
+                        This is a bug in deadlinks; please open an issue at https://github.com/deadlinks/cargo-deadlinks/issues/new/.");
+                }
                 process::exit(1);
             }
         };
@@ -85,20 +91,20 @@ fn main() {
 /// from the package name found there.
 ///
 /// All *.html files under the root directory will be checked.
-fn determine_dir() -> Vec<PathBuf> {
+fn determine_dir(no_build: bool) -> Vec<PathBuf> {
     let manifest = MetadataCommand::new()
         .no_deps()
         .exec()
         .unwrap_or_else(|err| {
             println!("error: {}", err);
             println!("help: if this is not a cargo directory, use `--dir`");
-            ::std::process::exit(1);
+            process::exit(1);
         });
     let doc = manifest.target_directory.join("doc");
 
     // originally written with this impressively bad jq query:
     // `.packages[] |select(.source == null) | .targets[] | select(.kind[] | contains("test") | not) | .name`
-    manifest
+    let paths = manifest
         .packages
         .into_iter()
         .filter(|package| package.source.is_none())
@@ -106,7 +112,18 @@ fn determine_dir() -> Vec<PathBuf> {
         .flatten()
         .filter(has_docs)
         .map(|target| doc.join(target.name.replace('-', "_")))
-        .collect()
+        .collect();
+
+    // Finally, build the documentation.
+    log::info!("building documentation using cargo");
+    let cargo = env::var("CARGO").expect(
+        "`cargo-deadlinks` must be run as either `cargo deadlinks` or with the `--dir` flag",
+    );
+    if !no_build && !Command::new(cargo).arg("doc").status().unwrap().success() {
+        process::exit(2);
+    } else {
+        paths
+    }
 }
 
 fn has_docs(target: &cargo_metadata::Target) -> bool {